all ok

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: cesar <cesar@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/08 09:33:59 by cefuente          #+#    #+#             */
-/*   Updated: 2023/11/17 11:57:34 by cesar            ###   ########.fr       */
+/*   Updated: 2023/11/20 08:27:49 by cesar            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
